+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Current
 
 * component `image`: new `max-height: 540px`
+* module `get-involved-list`: remove hands background by default (added a `hand` modifier if needed)
 * module `quick-access`: fixed video 1px bug on Firefox
 * module `rheader`:
   * added `target` for `locales`